% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dataset.R
\name{check_lot}
\alias{check_lot}
\title{Fish lot checker (AFB)}
\usage{
check_lot(
  id = NULL,
  type = NULL,
  min_size = NULL,
  max_size = NULL,
  nb = NULL,
  ind_measure = NULL,
  ind_id = NULL,
  ind_size = NULL,
  ...
)
}
\arguments{
\item{id}{int id of the lot}

\item{type}{character type of the lot (N, G, S/L, I)}

\item{min_size}{dbl minimum size of the lot}

\item{max_size}{dbl maximum size of the lot}

\item{nb}{int effectif of the lot}

\item{ind_measure}{data.frame individual measurement of the fish}

\item{ind_id}{variable name id of the lot in ind_measure}

\item{ind_size}{variable name for the size in ind_measure}
}
\value{
a list. Each element of the list correspond to a lot of a fishing
operation. If there is no problem in the lot, it returns NA. In the opposite
case, the function returns a character containing the type of error.
}
\description{
Check the quality of the lot
}
\details{
The function check if there are missing datas that can make the lot
invalid. The check are specific of each lot type. For lot "G", we check if
min or max size are missing, if min >= max size and the number of fish in the
lot is superior or equal to 10. For lot "S/L", we check if the number of measured
fishes is superior to 20 (standard protocol propose 30 be measured). In lot
"I", we check that the number of measured fishes correspond to the number of
fish recorded in the lot. For lot "N", we check if the number of measured
fish is equal to one.
}
