% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_summary_mod}
\alias{clean_summary_mod}
\title{Make summary table}
\usage{
clean_summary_mod(ci = NULL, mod = NULL, term_rp = NULL, group_rp = NULL)
}
\arguments{
\item{ci}{obj from bootMer}

\item{mod}{linear model}

\item{term_rp}{named character vector}

\item{group_rp}{named character vector}
}
\description{
Make summary table
}
