% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dataset.R
\name{get_check_lot}
\alias{get_check_lot}
\title{Main function for lot check}
\usage{
get_check_lot(
  lot = NULL,
  id_var = NULL,
  type_var = NULL,
  nb_var = NULL,
  min_var = NULL,
  max_var = NULL,
  species = NULL,
  measure = NULL,
  measure_id_var = NULL,
  size_var = NULL,
  future_enabled = FALSE,
  ...
)
}
\description{
return a data.frame with lot id, species and the
}
