% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dataset.R
\name{get_longest_consecutive_time_series}
\alias{get_longest_consecutive_time_series}
\title{Filter fish op to get op dataset without holes}
\usage{
get_longest_consecutive_time_series(x)
}
\description{
Filter fish op to get op dataset without holes
}
