% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dataset.R
\name{get_size_from_lot}
\alias{get_size_from_lot}
\title{Get size from lot (AFB)}
\usage{
get_size_from_lot(
  lot = NULL,
  id_var = NULL,
  type_var = NULL,
  nb_var = NULL,
  min_var = NULL,
  max_var = NULL,
  species = NULL,
  measure = NULL,
  measure_id_var = NULL,
  size_var = NULL,
  future_enabled = FALSE,
  ...
)
}
\arguments{
\item{lot}{data.frame}
}
\description{
Get size from lot (AFB)
}
