% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedness.R
\name{nestedness}
\alias{nestedness}
\title{Compute nestedness}
\usage{
nestedness(adj)
}
\arguments{
\item{adj}{an adjacency matrix}
}
\value{
double. The nestedness value
}
\description{
Compute the nestedness of a network according to Bastolla et al. (2009).
}
\details{
The nestedness is computed as follows: \eqn{\nu(A) = \frac{\sum
\nu_{i,j}}{\sum min(\nu_i, \nu_j)}}
Where
}
\references{
Bastolla, U., Fortuna, M. A., Pascual-García, A., Ferrera, A., Luque, B., & Bascompte, J. (2009). The architecture of mutualistic networks minimizes competition and increases biodiversity. Nature, 458(7241), 1018‑1020. https://doi.org/10.1038/nature07950
}
