% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_webs.R
\name{PlotWeb}
\alias{PlotWeb}
\title{plot web}
\usage{
PlotWeb(
  TL,
  webTL,
  colnode,
  abund,
  collink = "grey70",
  scale_abun = 0.01,
  rel_abun = TRUE,
  log_abun = FALSE
)
}
\description{
plot web
}
