% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_dataset.R
\name{rm_dbl_fishing_op}
\alias{rm_dbl_fishing_op}
\title{Remove fishing operation that are too close in time}
\usage{
rm_dbl_fishing_op(op = NULL, sep_threshold = 270, nb_sampling = 10)
}
\arguments{
\item{op}{data.frame generated by the vignette \code{aspe_database.Rmd}}

\item{sep_threshold}{int. Number of days minimum between two operations}

\item{nb_sampling}{int. Minimum number of the operation for a station to be
kept in the dataset}
}
\description{
Remove fishing operation that are too close in time
}
\details{
sep_threshold and nb_sampling are set to 270 and 10 by default
respectively.
}
