% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_webs.R
\name{set_color_species}
\alias{set_color_species}
\title{Colours species}
\usage{
set_color_species(
  node_list = NULL,
  species_list = NULL,
  resource_list = NULL,
  col_resource = NULL
)
}
\value{
a colour scale
}
\description{
Colours species
}
