% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeTrophicInteractions-package.R
\docType{package}
\name{sizeTrophicInteractions-package}
\alias{sizeTrophicInteractions}
\alias{sizeTrophicInteractions-package}
\title{sizeTrophicInteractions: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
