% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaweb_build.R
\name{sanatize_metaweb}
\alias{sanatize_metaweb}
\title{Sanatize fish length for metaweb}
\usage{
sanatize_metaweb(
  data = NULL,
  species = NULL,
  fish_diet_shift = NULL,
  nb_class = NULL
)
}
\description{
Sanatize fish length for metaweb
}
